"Add .gitignore file for various environments"
The .gitignore file was added to the project root to ignore files that should not be tracked by Git. This includes various development tools and environments such as IntelliJ, PyCharm, GoLand, WebStorm, Node, Go, Python, etc. This helps to keep the repository clean from unnecessary files that are created during the development process. This .gitignore was crafted to accommodate various use cases and development environments.